requirements.txt
pandas
pyarrow
numpy
tqdm

hydra-core
gurobipy

jupyterlab
ipywidgets
matplotlib
seaborn
plotly

# foreacsting
prophet

pytest
